// src/app/models/activity-detail.ts
export class ActivityDetail {
    added_date: string;
    address: string;
    attending: string;
    attending_count: string;
    capacity: string;
    category: string[] = [];
    city: string;
    company: string;
    country: string;
    cover: string;
    currency: string;
    description: string;
    email: string;
    end_date: string;
    end_time: string;
    event_id: string;
    fb_id: string;
    id: string;
    img_large: string;
    img_small: string;
    interested_count: string;
    is_wishlist: boolean;
    joined: string;
    latitude: string;
    link: string;
    location: string;
    location_id: string;
    longitude: string;
    maybe_count: string;
    mobile: string;
    name: string;
    owner_id: string;
    owner_name: string;
    owner_picture: string;
    phone: string;
    picture: string;
    place_id: string;
    price: string;
    score: string;
    slogan: string;
    source_link: string;
    space: string;
    start_date: string;
    start_time: string;
    timezone: string;
    title: string;
    user_id: string;
}